+++ src/components/center/ticket/Form/InputField.tsx
@@ -5,9 +5,10 @@ import { styled } from 'styled-components';
 import { SC } from '@styles/styles';
 
 type InputFieldProps = {
+  className?: 'required';
   label?: string;
   name: string;
-  value: string | number;
+  value: string | number | undefined;
   error?: boolean;
   unit?: string;
   placeholder?: string;
@@ -16,22 +17,30 @@ type InputFieldProps = {
   onChange: (e: ChangeEvent<HTMLInputElement>) => void;
 };
 
-export const InputField: FC<InputFieldProps> = ({ label, name, value, error = false, unit, ...rest }) => {
+export const InputField: FC<InputFieldProps> = ({ className, label, name, value, error = false, unit, ...rest }) => {
   return unit ? (
     <Unit>
-      {label && <SC.Label htmlFor={name}>{label}</SC.Label>}
+      {label && (
+        <SC.Label className={className || ''} htmlFor={name}>
+          {label}
+        </SC.Label>
+      )}
       <SC.InputField
         className={error ? 'error' : ''}
         id={name}
         name={name}
-        value={value === 0 ? '' : value}
+        value={value === 0 || value === undefined ? '' : value}
         {...rest}
       />
       <span className="unit">{unit}</span>
     </Unit>
   ) : (
     <div>
-      {label && <SC.Label htmlFor={name}>{label}</SC.Label>}
+      {label && (
+        <SC.Label className={className || ''} htmlFor={name}>
+          {label}
+        </SC.Label>
+      )}
       <SC.InputField
         className={error ? 'error' : ''}
         id={name}
